backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
python-multipart==0.0.9
python-dotenv==1.0.1
chromadb==0.5.5
openai==1.47.0
google-generativeai==0.8.1
pymupdf==1.24.10
httpx==0.27.2
pydantic==2.9.1
pydantic-settings==2.5.2
langchain-text-splitters==0.3.0
uuid6==2024.7.10
